build tweak
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@15588 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/switch_log.c
@@ -341,7 +341,7 @@ SWITCH_DECLARE(void) switch_log_vprintf(switch_text_channel_t channel, const cha
 #endif
 
 	if (level > 100) {
-		if (level - 100 > runtime.debug_level) {
+		if ((uint32_t)(level - 100) > runtime.debug_level) {
 			return;
 		}
